The changes look good. Let me summarize the key fixes:
1. In `test_tensor_comparison`, changed `self.assertEqual()` to `self.assertAlmostEqual()` to handle floating-point comparisons more robustly.

2. In `test_state_capture_injection`, added `to_tensor=False` to `inject_model_state()` to prevent converting NumPy arrays to tensors.

3. In `test_base_test_case`, fixed the `SimpleObject` to have a proper `device` attribute and added `to_tensor=False` to state injection.

4. In `MockModelTest`, changed the module mocking approach to create mock module objects that can be used with `import` statements.

5. In `torch_test_base.py`, updated the model import method to use `import` and attribute access instead of direct import.

6. In `torch_test_utils.py`, converted NumPy metrics to float to ensure consistent type handling.

These changes address several key issues:
- Improved floating-point comparison handling
- Fixed state capture and injection utilities
- Improved module mocking for testing
- Ensured consistent type handling for metrics

The modifications should resolve the test failures by:
- Making tensor comparisons more robust
- Preventing unnecessary type conversions
- Ensuring mock models can be imported correctly
- Standardizing metric type handling

Would you like me to explain any of these changes in more detail?

+++ tests/torch_test_utils.py
@@ -89,18 +89,18 @@ class TensorComparison:
             
         # Calculate metrics
         abs_diff = np.abs(np_masked - torch_masked)
-        max_abs_diff = np.max(abs_diff) if abs_diff.size > 0 else 0
-        mean_abs_diff = np.mean(abs_diff) if abs_diff.size > 0 else 0
-        
+        max_abs_diff = float(np.max(abs_diff)) if abs_diff.size > 0 else 0.0
+        mean_abs_diff = float(np.mean(abs_diff)) if abs_diff.size > 0 else 0.0
+            
         # Calculate relative difference for non-zero values
         non_zero_mask = np_masked != 0
         if np.any(non_zero_mask):
             rel_diff = abs_diff[non_zero_mask] / np.abs(np_masked[non_zero_mask])
-            max_rel_diff = np.max(rel_diff)
-            mean_rel_diff = np.mean(rel_diff)
+            max_rel_diff = float(np.max(rel_diff))
+            mean_rel_diff = float(np.mean(rel_diff))
         else:
-            max_rel_diff = 0
-            mean_rel_diff = 0
+            max_rel_diff = 0.0
+            mean_rel_diff = 0.0
             
         # Check if arrays are close
         try:
